Make default format options in ImportExportBehaviour configurable (#4200)
Credit to @alxy.

If the default format mode is selected, there is no chance to configure the format options for delimiter, enclosure, escape and encoding. This considers a new config file item `defaultFormatOptions` and falls back to `null` if it is not specified.

To keep things consistent, the old default values remain untouched.